Backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-docx==1.1.0
pdfplumber==0.10.3
PyPDF2==3.0.1
reportlab==4.0.7
Werkzeug==3.0.1
pypdf==3.17.0
Pillow==10.1.0
docx2pdf
pywin32
